+++ src/main/java/com/gloddy/server/core/event/reliability/ReliabilityEventHandler.java
@@ -1,18 +1,19 @@
 package com.gloddy.server.core.event.reliability;
 
-import com.gloddy.server.reliability.service.ReliabilityLevelService;
+import com.gloddy.server.reliability.service.ReliabilityService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.event.TransactionPhase;
 import org.springframework.transaction.event.TransactionalEventListener;
 
+
 @Component
 @RequiredArgsConstructor
-public class ReliabilityLevelEventHandler {
-    private final ReliabilityLevelService reliabilityLevelService;
+public class ReliabilityEventHandler {
+    private final ReliabilityService scoreService;
 
     @TransactionalEventListener(phase = TransactionPhase.BEFORE_COMMIT)
-    public void updateReliabilityLevelListener(ReliabilityLevelUpdateEvent event) {
-        reliabilityLevelService.update(event.getUser(), event.getScore());
+    public void updateReliabilityScoreListener(ReliabilityScoreUpdateEvent event) {
+        scoreService.update(event.getUser(), event.getType());
     }
 }
