+++ src/main/java/com/gloddy/server/core/event/reliability/ReliabilityScoreUpdateEvent.java
@@ -1,4 +1,4 @@
-package com.gloddy.server.core.event.score;
+package com.gloddy.server.core.event.reliability;
 
 import com.gloddy.server.auth.entity.User;
 import com.gloddy.server.reliability.entity.vo.ScoreType;
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-public class ScoreUpdateEvent {
+public class ReliabilityScoreUpdateEvent {
     private User user;
     private ScoreType type;
 }
